Fixing formatting for the image constructor tests
